Bugfixes for CPU rendering
Fix segfault caused by RectOutlineRenderer not properly checking bounds on pixel array.
Fix positioning of RectOutline sides in other locations when they were off the screen.

When you zoom in too far, an integer overflow will occur and the CPU rendered objects disappear.
This occured at widths of about 1e-7 (for our test pattern). Improved for now by int -> int64_t

Later need to replace everything except the actual array access with Real's I think. Then make Real's arbitrary precision. Done.

There is a slight size difference between CPU rendered and GPU rendered objects; not sure what the source of this is. Blame rounding.

GPU sometimes leaves off the corner pixels of the rectangles by the way.

+++ src/objectrenderer.h
@@ -44,12 +44,12 @@ namespace IPDF
 			struct CPURenderTarget
 			{
 				uint8_t * pixels;
-				int w;
-				int h;
+				int64_t w;
+				int64_t h;
 			};
 			struct CPURenderBounds
 			{
-				int x; int y; int w; int h;
+				int64_t x; int64_t y; int64_t w; int64_t h;
 				CPURenderBounds(const Rect & bounds, const View & view, const CPURenderTarget & target);
 			};
